+++ .gitignore
@@ -4,3 +4,4 @@ dist
 .DS_Store
 *.local
 pnpm-lock.yaml
+.idea
