+++ src/parser/index.ts
@@ -1,0 +1,14 @@
+import JSONToHTML from "./JSONToHTML";
+import textToHTML from "./textToHTML";
+import textToJSON from "./textToJSON";
+import { jsonRender, autoToHTML } from "./autoToHTML";
+
+
+
+export {
+  JSONToHTML,
+  textToHTML,
+  textToJSON,
+  jsonRender,
+  autoToHTML,
+};
