+++ src/utils.ts
@@ -6,16 +6,21 @@
 
 import { motdJsonType } from "./types";
 
+
+
+
 // 類型檢查
 export function isMotdJSONType(object: object): object is motdJsonType {
   return !!object;
-}
+};
+
+
 
 /**
  *
- * #### replace all html tags to &...
+ * replace all html tags to &...
  */
-export const htmlStringFormatting = (text: string): string => {
+export function htmlStringFormatting(text: string): string {
   return (
     text
       // space
@@ -30,4 +35,36 @@ export const htmlStringFormatting = (text: string): string => {
       .replace(/\n/g, "<br/>")
   );
   // .replace(/\//g, '&#x2F;');
+};
+
+
+
+/**
+ * clean html tags
+ * 
+ * @param text
+ * example: `<span>hello world</span>`
+ * 
+ * result: `hello world`
+ */
+export function cleanHtmlTags(text: string): string {
+  return text.replace(/<[^>]*>?/gm, '');
+};
+
+
+
+/**
+ * clean motd tags
+ * 
+ * Clean all tags from motd source string.
+ * 
+ * @param {string} text - motd string include tag § will remove
+ */
+export function cleanTags(text: string) {
+  const REGEX = /(?:§)([0-9a-fA-FklmnorFKLMNOR])/g;
+  let textResult = "";
+
+  textResult = text.replace(REGEX, "");
+
+  return textResult;
 };
